feat: add first prototype of scripting in rust
This is a foundation that allows a dev to write a list of commands that is applied on a saved show file to handle repetitive tasks.
The script is controlled by a (currently compiled into the program) yaml file.

This feature should be moved into the application, probably use another language (like python, rhai, lua?) and instead of providing a yaml file it should allow the user to pick elements (like a group) via the ui.
As this requires some major features in the ui layer this scripting crate provides a good middle ground for myself.